Authentication using Passport

+++ controllers/indexController.js
@@ -1,41 +1,50 @@
-import mongoose from 'mongoose'
-import {join} from 'path'
 
 import StudentModel from '../models/student.js'
 
 class IndexController{
-    static isLogged = false
-    static allstudents = async(req, res, isLogged) => {
-        try{
-            this.isLogged = isLogged
+
+    // Show all Students
+    static allstudents = async(req, res) => {
+      
+       if (req.isAuthenticated()){
+        try{ 
             const allData = await StudentModel.find()
-            res.render("index.ejs",{allData:allData,isLogged:this.isLogged })
+            res.render("index.ejs",{allData:allData })
 
         }catch(error){
             console.log(error)
         }
+       }
+       else{
+        res.redirect('/')
+       }
         
     }
+
+    // Add a new Student
     static addstudent = async(req, res) => {
-        try{
-            const {name,age,fees}=req.body
-            const StudentDoc = new StudentModel({
-                name:name,
-                age:age,
-                fees:fees
-,
-            })
-            await StudentDoc.save()
-            this.allstudents(req,res)
-
-        }catch(err){
-
-        }
-        
+        if (req.isAuthenticated()){
+            try{
+                const {name,age,fees}=req.body
+                const StudentDoc = new StudentModel({
+                    name:name,
+                    age:age,
+                    fees:fees
+                })
+                await StudentDoc.save()
+                this.allstudents(req,res)
+    
+            }catch(err){
+                console.log(err)
+            }
+        }else{
+            res.redirect('/')
+        }               
     }
+
+    // Render the update page for updating the student's details
     static editstudent = async(req,res) =>{
-        console.log(this.isLogged)
-        if (this.isLogged){
+        if (req.isAuthenticated()){
             try{
                 const id = req.params.id
                  const data = await StudentModel.findById(id)
@@ -43,31 +52,42 @@ class IndexController{
              }catch(error){
                  console.log(error)
              }
-            
+        }else{
+            res.redirect('/')
+        }    
+    }
+
+    // Update the student details
+    static updateStudent = async(req,res)=>{
+        if (req.isAuthenticated()){
+            try{
+                const id = req.params.id
+                const updatedData = req.body
+                const result = await StudentModel.updateOne({_id:id},updatedData)
+                res.redirect('/student/')
+            }catch(error){
+                console.log(error)
+            }
+        }else{
+            res.redirect('/')
         }
-    
         
     }
-    static updateStudent = async(req,res)=>{
-        try{
-            const id = req.params.id
-            const updatedData = req.body
-            const result = await StudentModel.updateOne({_id:id},updatedData)
-            console.log(result)
-            res.redirect('/student/')
-        }catch(error){
-            console.log(error)
-        }
-    }
+
+    // Delete the student from the Database
     static deleteStudent = async(req,res)=>{
-        try{
-            const id = req.params.id
-            const result = await StudentModel.deleteOne({_id:id})
-            console.log(result)
-            res.redirect('/student/')
-        }catch(error){
-            console.log(error)
+        if (req.isAuthenticated()){
+            try{
+                const id = req.params.id
+                const result = await StudentModel.deleteOne({_id:id})
+                res.redirect('/student/')
+            }catch(error){
+                console.log(error)
+            }
+        }else{
+            res.redirect('/')
         }
+        
     }
 }
